requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.7.4
dnspython==2.6.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pymongo==4.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
yarl==1.9.4
